pyproject.toml
[project]
name = "final-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-core>=1.0.2",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "matplotlib>=3.7.0",
    "mcp[cli]>=1.20.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.0",
    "streamlit-chat>=0.1.1",
]
